+++ plugins/home/src/index.ts
@@ -32,8 +32,11 @@ export {
   ComponentTab,
   WelcomeTitle,
   HeaderWorldClock,
+  HomePageTopVisited,
+  HomePageRecentlyVisited,
 } from './plugin';
 export * from './components';
 export * from './assets';
 export * from './homePageComponents';
 export * from './deprecated';
+export * from './api';