+++ packages/cli-node/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/cli-node
 
+## 0.1.5-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/cli-common@0.1.13-next.0
+  - @backstage/errors@1.2.2
+  - @backstage/types@1.1.1
+
 ## 0.1.4
 
 ### Patch Changes