+++ plugins/auth-backend-module-google-provider/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @backstage/plugin-auth-backend-module-google-provider
 
+## 0.1.2-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.5-next.1
+  - @backstage/plugin-auth-node@0.3.2-next.1
+
+## 0.1.2-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-auth-node@0.3.2-next.0
+  - @backstage/backend-plugin-api@0.6.5-next.0
+
 ## 0.1.0
 
 ### Minor Changes